requirements.txt
numpy
scipy
pandas
bokeh
colorcet
PyYAML